requirements.txt
beautifulsoup4==4.11.1
requests==2.28.1
selenium==4.4.3
webdriver-manager==3.8.3
python-dotenv==0.20.0